refactor: use u32/u16 for StreamConfig, remove from_signed
- Change StreamConfig::sample_rate_hz from usize to u32
- Change StreamConfig::num_channels from usize to u16
- num_frames() and num_samples() still return usize (buffer sizing)
- Remove from_signed() constructor and sample_rate_hz_signed() accessor
- Remove FormatValidity::InvalidRate variant (negative rates impossible with u32)
- Update MaxProcessingRate::as_hz() to return u32
- Update suitable_process_rate/sample_rate_supports_multi_band to use u32
- Add as usize casts at internal boundaries (AudioBuffer, AudioConverter)
- FFI layer handles i32 -> u32/u16 conversion with clamping
- Update all tests, benchmarks, and integration tests

+++ crates/sonora/src/audio_buffer.rs
@@ -264,7 +264,10 @@ impl AudioBuffer {
         stream_config: &StreamConfig,
     ) {
         debug_assert_eq!(stream_config.num_frames(), self.input_num_frames);
-        debug_assert_eq!(stream_config.num_channels(), self.input_num_channels);
+        debug_assert_eq!(
+            stream_config.num_channels() as usize,
+            self.input_num_channels
+        );
         self.restore_num_channels();
 
         let downmix_needed = self.input_num_channels > 1 && self.num_channels == 1;
@@ -344,9 +347,9 @@ impl AudioBuffer {
         }
 
         // Copy channel 0 to any extra output channels.
-        if self.num_channels < stream_config.num_channels() {
+        if self.num_channels < stream_config.num_channels() as usize {
             let (first, rest) = stacked_data.split_at_mut(1);
-            for i in self.num_channels..stream_config.num_channels() {
+            for i in self.num_channels..stream_config.num_channels() as usize {
                 rest[i - 1][..self.output_num_frames]
                     .copy_from_slice(&first[0][..self.output_num_frames]);
             }
@@ -393,7 +396,10 @@ impl AudioBuffer {
         interleaved_data: &[i16],
         stream_config: &StreamConfig,
     ) {
-        debug_assert_eq!(stream_config.num_channels(), self.input_num_channels);
+        debug_assert_eq!(
+            stream_config.num_channels() as usize,
+            self.input_num_channels
+        );
         debug_assert_eq!(stream_config.num_frames(), self.input_num_frames);
         self.restore_num_channels();
 
@@ -478,7 +484,7 @@ impl AudioBuffer {
         stream_config: &StreamConfig,
         interleaved_data: &mut [i16],
     ) {
-        let config_num_channels = stream_config.num_channels();
+        let config_num_channels = stream_config.num_channels() as usize;
         debug_assert!(config_num_channels == self.num_channels || self.num_channels == 1);
         debug_assert_eq!(stream_config.num_frames(), self.output_num_frames);
